rag/requirements.txt
google-generativeai
langchain-google-genai
chromadb
pypdf
Pillow
langchain
langchain-community
pandas